[조유태] 분기별 -> 월별, 이슈 수정

+++ voyage/src/main/java/com/gdu/voyage/vo/AccomPayment.java
@@ -9,7 +9,7 @@ public class AccomPayment {
 	private String memberId;
 	private String accomCheckIn;
 	private String accomCheckOut;
-	private int accomAmount;
+	private long accomAmount;
 	private int accomUsePerson;
 	private String accomPaymentState;
 	private String receipt;
